Add `--ignore-bind`
Workaround for issue 65.

This flag disables the acceleration for `bind` syscalls,
as the current hook is not robust enough for `bind` with complicated setup.

Anyway, the performance of `bind` is not really problematic so far in the plain Rootless Containers.
So we can just keep enabling bypass4netns for `connect` and relevant syscalls, and call it for a day.

See the logs of RootlessKit CI:
https://github.com/rootless-containers/rootlesskit/actions/runs/8558692577/job/23453781376
- Benchmark: TCP Ports (network driver=slirp4netns, port driver=builtin): 34.8 Gbps, 34.7 Gbps
- Benchmark: TCP Ports (network driver=pasta, port driver=implicit):      39.3 Gbps, 39.2 Gbps
- Benchmark: UDP Ports (network driver=slirp4netns, port driver=builtin): 25.1 Gbps, 17.2 Gbps
- Benchmark: UDP Ports (network driver=pasta, port driver=implicit):      37.9 Gbps, 13.2 Gbps

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>